requirements.txt
aiohttp==3.11.10
cryptg==0.5.0.post0
python-dotenv==1.0.1
Telethon==1.38.1
websockets==14.1